Registration service: apply naming for packages
