+++ registration/src/main/java/com/kyriba/conference/registration/api/dto/ChangeTicketOwnerRequest.java
@@ -1,4 +1,4 @@
-package com.kyriba.registration.api.dto;
+package com.kyriba.conference.registration.api.dto;
 
 
 import io.swagger.annotations.ApiModel;
