+++ registration/src/test/java/com/kyriba/conference/registration/RegistrationApplicationTests.java
@@ -1,4 +1,4 @@
-package com.kyriba.registration;
+package com.kyriba.conference.registration;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
